refactor: move connection dialogs to connection list
